requirements.txt
pyyaml>=6.0
faker>=18.0
numpy>=1.24
python-dateutil>=2.8
click>=8.0
pydantic>=2.0